Move serial tests to separate files
