+++ test/snapshot-workflow/removing-all-snapshots.js
@@ -1,0 +1,29 @@
+const test = require('@ava/test');
+const exec = require('../helpers/exec');
+const {beforeAndAfter} = require('./helpers/macros');
+
+test.serial(
+	'Removing all snapshots from a test retains its data',
+	beforeAndAfter,
+	{
+		cwd: exec.cwd('removing-all-snapshots')
+	},
+	async (t, {before, after}) => {
+		t.deepEqual(after.snapshot, before.snapshot);
+		t.is(after.report, before.report);
+	}
+);
+
+test.serial(
+	'With --update-snapshots, removing all snapshots from a test removes the block',
+	beforeAndAfter,
+	{
+		cwd: exec.cwd('removing-all-snapshots'),
+		after: {cli: ['--update-snapshots']}
+	},
+	async (t, {before, after}) => {
+		t.notDeepEqual(after.snapshot, before.snapshot);
+		t.not(after.report, before.report);
+		t.snapshot(after.report, 'snapshot report after removing all snapshots from a test');
+	}
+);
