+++ test/snapshot-workflow/reorder.js
@@ -1,0 +1,29 @@
+const test = require('@ava/test');
+const exec = require('../helpers/exec');
+const {beforeAndAfter} = require('./helpers/macros');
+
+test.serial(
+	'Reordering tests does not change the .snap or .md',
+	beforeAndAfter,
+	{
+		cwd: exec.cwd('reorder')
+	},
+	async (t, {before, after}) => {
+		t.deepEqual(after.snapshot, before.snapshot);
+		t.is(after.report, before.report);
+	}
+);
+
+test.serial(
+	'With --update-snapshots, reordering tests reorders the .snap and .md',
+	beforeAndAfter,
+	{
+		cwd: exec.cwd('reorder'),
+		after: {cli: ['--update-snapshots']}
+	},
+	async (t, {before, after}) => {
+		t.notDeepEqual(after.snapshot, before.snapshot);
+		t.not(after.report, before.report);
+		t.snapshot(after.report, 'snapshot report after reordering tests');
+	}
+);
