+++ test/snapshot-workflow/snapshots/changing-label.js.md
@@ -1,0 +1,24 @@
+# Snapshot report for `test/snapshot-workflow/changing-label.js`
+
+The actual snapshot is saved in `changing-label.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## With --update-snapshots, changing a snapshot's label updates the .snap and .md
+
+> snapshot report after changing a label
+
+    `# Snapshot report for `test.js`␊
+    ␊
+    The actual snapshot is saved in `test.js.snap`.␊
+    ␊
+    Generated by [AVA](https://avajs.dev).␊
+    ␊
+    ## foo␊
+    ␊
+    > a new message␊
+    ␊
+        {␊
+          foo: 'one',␊
+        }␊
+    `
