+++ test/snapshot-workflow/snapshots/removing-all-snapshots.js.md
@@ -1,0 +1,24 @@
+# Snapshot report for `test/snapshot-workflow/removing-all-snapshots.js`
+
+The actual snapshot is saved in `removing-all-snapshots.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## With --update-snapshots, removing all snapshots from a test removes the block
+
+> snapshot report after removing all snapshots from a test
+
+    `# Snapshot report for `test.js`␊
+    ␊
+    The actual snapshot is saved in `test.js.snap`.␊
+    ␊
+    Generated by [AVA](https://avajs.dev).␊
+    ␊
+    ## bar␊
+    ␊
+    > Snapshot 1␊
+    ␊
+        {␊
+          bar: 'one',␊
+        }␊
+    `
