+++ test/snapshot-workflow/snapshots/removing-snapshots.js.md
@@ -1,0 +1,24 @@
+# Snapshot report for `test/snapshot-workflow/removing-snapshots.js`
+
+The actual snapshot is saved in `removing-snapshots.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## With --update-snapshots, removing a snapshot assertion removes its data
+
+> snapshot report after removing a snapshot
+
+    `# Snapshot report for `test.js`␊
+    ␊
+    The actual snapshot is saved in `test.js.snap`.␊
+    ␊
+    Generated by [AVA](https://avajs.dev).␊
+    ␊
+    ## foo␊
+    ␊
+    > Snapshot 1␊
+    ␊
+        {␊
+          foo: 'one',␊
+        }␊
+    `
