+++ test/snapshot-workflow/snapshots/removing-test.js.md
@@ -1,0 +1,24 @@
+# Snapshot report for `test/snapshot-workflow/removing-test.js`
+
+The actual snapshot is saved in `removing-test.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## With --update-snapshots, removing a test removes its block
+
+> snapshot report after removing a test
+
+    `# Snapshot report for `test.js`␊
+    ␊
+    The actual snapshot is saved in `test.js.snap`.␊
+    ␊
+    Generated by [AVA](https://avajs.dev).␊
+    ␊
+    ## foo␊
+    ␊
+    > Snapshot 1␊
+    ␊
+        {␊
+          foo: 'one',␊
+        }␊
+    `
