+++ test/snapshot-workflow/snapshots/reorder.js.md
@@ -1,0 +1,32 @@
+# Snapshot report for `test/snapshot-workflow/reorder.js`
+
+The actual snapshot is saved in `reorder.js.snap`.
+
+Generated by [AVA](https://avajs.dev).
+
+## With --update-snapshots, reordering tests reorders the .snap and .md
+
+> snapshot report after reordering tests
+
+    `# Snapshot report for `test.js`␊
+    ␊
+    The actual snapshot is saved in `test.js.snap`.␊
+    ␊
+    Generated by [AVA](https://avajs.dev).␊
+    ␊
+    ## second test␊
+    ␊
+    > Snapshot 1␊
+    ␊
+        {␊
+          bar: 'baz',␊
+        }␊
+    ␊
+    ## first test␊
+    ␊
+    > Snapshot 1␊
+    ␊
+        {␊
+          foo: 'bar',␊
+        }␊
+    `
