+++ test/snapshot-workflow/snapshots/test.js.md
@@ -39,7 +39,7 @@ Generated by [AVA](https://avajs.dev).
 
 ## Adding more snapshots to a test adds them to the .snap and .md
 
-> snapshort report after adding snapshots
+> snapshot report after adding snapshots
 
     `# Snapshot report for `test.js`␊
     ␊
@@ -89,25 +89,6 @@ Generated by [AVA](https://avajs.dev).
         }␊
     `
 
-## With --update-snapshots, changing a snapshot's label updates the .snap and .md
-
-> snapshot report after changing a label
-
-    `# Snapshot report for `test.js`␊
-    ␊
-    The actual snapshot is saved in `test.js.snap`.␊
-    ␊
-    Generated by [AVA](https://avajs.dev).␊
-    ␊
-    ## foo␊
-    ␊
-    > a new message␊
-    ␊
-        {␊
-          foo: 'one',␊
-        }␊
-    `
-
 ## Changing a test's title adds a new block, puts the old block at the end
 
 > snapshot report after changing a title
@@ -125,84 +106,8 @@ Generated by [AVA](https://avajs.dev).
         {␊
           foo: 'one',␊
         }␊
-    `
-
-## With --update-snapshots, reordering tests reorders the .snap and .md
-
-> snapshot report after reordering tests
-
-    `# Snapshot report for `test.js`␊
     ␊
-    The actual snapshot is saved in `test.js.snap`.␊
-    ␊
-    Generated by [AVA](https://avajs.dev).␊
-    ␊
-    ## second test␊
-    ␊
-    > Snapshot 1␊
-    ␊
-        {␊
-          bar: 'baz',␊
-        }␊
-    ␊
-    ## first test␊
-    ␊
-    > Snapshot 1␊
-    ␊
-        {␊
-          foo: 'bar',␊
-        }␊
-    `
-
-## With --update-snapshots, removing a snapshot assertion removes its data
-
-> snapshot report after removing a snapshot
-
-    `# Snapshot report for `test.js`␊
-    ␊
-    The actual snapshot is saved in `test.js.snap`.␊
-    ␊
-    Generated by [AVA](https://avajs.dev).␊
-    ␊
-    ## foo␊
-    ␊
-    > Snapshot 1␊
-    ␊
-        {␊
-          foo: 'one',␊
-        }␊
-    `
-
-## With --update-snapshots, removing all snapshots from a test removes the block
-
-> snapshot report after removing all snapshots from a test
-
-    `# Snapshot report for `test.js`␊
-    ␊
-    The actual snapshot is saved in `test.js.snap`.␊
-    ␊
-    Generated by [AVA](https://avajs.dev).␊
-    ␊
-    ## bar␊
-    ␊
-    > Snapshot 1␊
-    ␊
-        {␊
-          bar: 'one',␊
-        }␊
-    `
-
-## With --update-snapshots, removing a test removes its block
-
-> snapshot report after removing a test
-
-    `# Snapshot report for `test.js`␊
-    ␊
-    The actual snapshot is saved in `test.js.snap`.␊
-    ␊
-    Generated by [AVA](https://avajs.dev).␊
-    ␊
-    ## foo␊
+    ## a title␊
     ␊
     > Snapshot 1␊
     ␊
